Replace the original api fetching method with <getServerSideProps>
