--- goplus.org/pages/download/README.md
@@ -1,5 +1,0 @@
-```shell
-node proxy_server.js
-```
-
-Before running the code of the website, run the above script first.
