--- goplus.org/pages/download/proxy_server.js
@@ -1,17 +1,0 @@
-const { createProxyMiddleware } = require('http-proxy-middleware');
-const express = require('express');
-
-const app = express();
-
-app.use('/github', createProxyMiddleware({
-  target: 'https://api.github.com',
-  changeOrigin: true,
-  pathRewrite: {
-    '^/github': ''
-  },
-}));
-
-const PORT = 3001;
-app.listen(PORT, () => {
-  console.log(`Proxy server is running on http://localhost:${PORT}`);
-});
